Creating Test Release (#32)
* Kind of manual but get the tags working. (#28)

This has a flaw in that the move from develop -> Test could be kind of borked.  Should be using SHAs but since we are dealing with small teams this should be good enough in that we can put out a "PR HOLD" or something like that.

Co-authored-by: Chris Jimison <chris@roostership.com>

* 30 update build (#31)

* Kind of manual but get the tags working. (#28) (#29)

This has a flaw in that the move from develop -> Test could be kind of borked.  Should be using SHAs but since we are dealing with small teams this should be good enough in that we can put out a "PR HOLD" or something like that.

Co-authored-by: Chris Jimison <chris@roostership.com>

* Trying to fix issue dirctly.

* Yet another test

* Updated the paths for docker

* This should close it out

* Updated build scripts to use the new image tags.

---------

Co-authored-by: Chris Jimison <chris@roostership.com>

---------

Co-authored-by: Chris Jimison <chris@roostership.com>